+++ src/Convert/Public/ConvertFrom-EscapedUrl.ps1
@@ -1,0 +1,14 @@
+function ConvertFrom-EscapedUrl {
+    [CmdletBinding()]
+    param (
+        [Parameter(ValueFromPipeline, ValueFromPipelineByPropertyName)]
+        [ValidateNotNullOrEmpty()]
+        [string[]]$Url
+    )
+
+    process {
+        foreach ($u in $Url) {
+            [System.Uri]::UnescapeDataString($u)
+        }
+    }
+}