+++ src/Convert/Public/ConvertTo-EscapedUrl.ps1
@@ -1,0 +1,14 @@
+function ConvertTo-EscapedUrl {
+    [CmdletBinding()]
+    param (
+        [Parameter(ValueFromPipeline, ValueFromPipelineByPropertyName)]
+        [ValidateNotNullOrEmpty()]
+        [string]$Url
+    )
+
+    process {
+        foreach ($u in $Url) {
+            [System.Uri]::EscapeDataString($u)
+        }
+    }
+}